chatbot-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai==1.50.0
qdrant-client==1.11.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.0
httpx==0.27.0
tiktoken==0.7.0
